Refactor to isolate Database objects to its own file
