+++ scripts/lava.py
@@ -1,6 +1,5 @@
 import math
 import os
-import random
 import shlex
 import struct
 import subprocess
@@ -12,364 +11,22 @@ from os.path import join
 from subprocess import PIPE
 from subprocess import check_call
 
-from sqlalchemy import Column
-from sqlalchemy import ForeignKey
-from sqlalchemy import Table
-from sqlalchemy import create_engine
-from sqlalchemy.dialects import postgresql
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import load_only
-from sqlalchemy.orm import relationship
-from sqlalchemy.orm import sessionmaker
-from sqlalchemy.sql.expression import func
-from sqlalchemy.types import BigInteger
-from sqlalchemy.types import Boolean
-from sqlalchemy.types import Float
-from sqlalchemy.types import Integer
-from sqlalchemy.types import Text
 
-from composite import Composite
 from process_compile_commands import get_c_files
 from process_compile_commands import process_compile_commands
 from test_crash import process_crash
 from dotenv import load_dotenv
+from database_types import Bug, DuaBytes, AttackPoint, Build, Run
 
 load_dotenv()
 
-Base = declarative_base()
+
 
 debugging = False
 NUM_BUGTYPES = 3  # Make sure this matches what's in lavaTool
 
 
-class Loc(Composite):
-    column = Integer
-    line = Integer
 
-
-class ASTLoc(Composite):
-    filename = Text
-    begin = Loc
-    end = Loc
-
-
-class Range(Composite):
-    low = Integer
-    high = Integer
-
-
-class SourceLval(Base):
-    __tablename__ = 'sourcelval'
-
-    id = Column(Integer, primary_key=True)
-    loc = ASTLoc.composite('loc')
-    ast_name = Column(Text)
-
-    def __str__(self):
-        return 'Lval[{}](loc={}:{}, ast="{}")'.format(
-            self.id, self.loc.filename, self.loc.begin.line, self.ast_name
-        )
-
-
-class LabelSet(Base):
-    __tablename__ = 'labelset'
-
-    id = Column(BigInteger, primary_key=True)
-    ptr = Column(BigInteger)
-    inputfile = Column(Text)
-    labels = Column(postgresql.ARRAY(Integer))
-
-    def __repr__(self):
-        return str(self.labels)
-
-
-dua_viable_bytes = \
-    Table('dua_viable_bytes', Base.metadata,
-          Column('object_id', BigInteger, ForeignKey('dua.id')),
-          Column('index', BigInteger),
-          Column('value', BigInteger, ForeignKey('labelset.id')))
-
-
-class Dua(Base):
-    __tablename__ = 'dua'
-
-    id = Column(BigInteger, primary_key=True)
-    lval_id = Column('lval', BigInteger, ForeignKey('sourcelval.id'))
-    all_labels = Column(postgresql.ARRAY(Integer))
-    inputfile = Column(Text)
-    max_tcn = Column(Integer)
-    max_cardinality = Column(Integer)
-    instr = Column(BigInteger)
-    fake_dua = Column(Boolean)
-
-    lval = relationship("SourceLval")
-    viable_bytes = relationship("LabelSet", secondary=dua_viable_bytes)
-
-    def __str__(self):
-        return 'DUA[{}](lval={}, labels={}, viable={}, \
-                input={}, instr={}, fake_dua={})'.format(
-            self.id, self.lval, self.all_labels, self.viable_bytes,
-            self.inputfile, self.instr, self.fake_dua
-        )
-
-
-class DuaBytes(Base):
-    __tablename__ = 'duabytes'
-
-    id = Column(BigInteger, primary_key=True)
-    dua_id = Column('dua', BigInteger, ForeignKey('dua.id'))
-    selected = Range.composite('selected')
-    all_labels = Column(postgresql.ARRAY(Integer))
-
-    dua = relationship("Dua")
-
-    def __str__(self):
-        return 'DUABytes[DUA[{}:{}, {}, {}]][{}:{}](labels={})'.format(
-            self.dua.lval.loc.filename, self.dua.lval.loc.begin.line,
-            self.dua.lval.ast_name, 'fake' if self.dua.fake_dua else 'real',
-            self.selected.low, self.selected.high, self.all_labels)
-
-
-class AttackPoint(Base):
-    __tablename__ = 'attackpoint'
-
-    id = Column(BigInteger, primary_key=True)
-    loc = ASTLoc.composite('loc')
-    typ = Column('type', Integer)
-
-    # enum Type {
-    FUNCTION_CALL = 0
-    POINTER_READ = 1
-    POINTER_WRITE = 2
-    QUERY_POINT = 3
-    PRINTF_LEAK = 4
-
-    # } type;
-
-    def __str__(self):
-        type_strs = [
-            "ATP_FUNCTION_CALL",
-            "ATP_POINTER_READ",
-            "ATP_POINTER_WRITE",
-            "ATP_QUERY_POINT",
-            "ATP_PRINTF_LEAK",
-            "ATP_MALLOC_OFF_BY_ONE"
-        ]
-        return 'ATP[{}](loc={}:{}, type={})'.format(
-            self.id, self.loc.filename, self.loc.begin.line, type_strs[self.typ]
-        )
-
-
-build_bugs = \
-    Table('build_bugs', Base.metadata,
-          Column('object_id', BigInteger, ForeignKey('build.id')),
-          Column('index', BigInteger, default=0),
-          Column('value', BigInteger, ForeignKey('bug.id')))
-
-
-class Bug(Base):
-    __tablename__ = 'bug'
-
-    # enum Type {
-    PTR_ADD = 0
-    RET_BUFFER = 1
-    REL_WRITE = 2
-    PRINTF_LEAK = 3
-    MALLOC_OFF_BY_ONE = 4
-    # };
-    type_strings = ['BUG_PTR_ADD', 'BUG_RET_BUFFER',
-                    'BUG_REL_WRITE', 'BUG_PRINTF_LEAK', 'MALLOC_OFF_BY_ONE']
-
-    id = Column(BigInteger, primary_key=True)
-    type = Column(Integer)
-    trigger_id = Column('trigger', BigInteger, ForeignKey('duabytes.id'))
-    trigger_lval_id = Column('trigger_lval', BigInteger,
-                             ForeignKey('sourcelval.id'))
-    atp_id = Column('atp', BigInteger, ForeignKey('attackpoint.id'))
-
-    trigger = relationship("DuaBytes")
-    trigger_lval = relationship("SourceLval")
-
-    max_liveness = Column(Float)
-    magic = Column(Integer)
-
-    atp = relationship("AttackPoint")
-
-    extra_duas = Column(postgresql.ARRAY(BigInteger))
-
-    builds = relationship("Build", secondary=build_bugs,
-                          back_populates="bugs")
-
-    def __str__(self):
-        return 'Bug[{}](type={}, trigger={}, atp={})'.format(
-            self.id, Bug.type_strings[self.type], self.trigger, self.atp)
-
-
-class Build(Base):
-    __tablename__ = 'build'
-
-    id = Column(BigInteger, primary_key=True)
-    compile = Column(Boolean)
-    output = Column(Text)
-
-    bugs = relationship("Bug", secondary=build_bugs,
-                        back_populates="builds")
-
-
-class Run(Base):
-    __tablename__ = 'run'
-
-    id = Column(BigInteger, primary_key=True)
-    build_id = Column('build', BigInteger, ForeignKey('build.id'))
-    fuzzed_id = Column('fuzzed', BigInteger, ForeignKey('bug.id'))
-    exitcode = Column(Integer)
-    output = Column(Text)
-    success = Column(Boolean)
-    validated = Column(Boolean)
-
-    build = relationship("Build")
-    fuzzed = relationship("Bug")
-
-
-class LavaDatabase(object):
-    def __init__(self, project):
-        self.project = project
-        self.engine = create_engine(
-            "postgresql+psycopg2://{}@{}:{}/{}".format(
-                project['database_user'],
-                project['database'],
-                project['database_port'],
-                project['db']
-            )
-        )
-        self.Session = sessionmaker(bind=self.engine)
-        self.session = self.Session()
-
-    # If we have over a million bugs, don't bother counting things
-    def huge(self):
-        return self.session.query(Bug.id).count() > 1000000
-
-    def uninjected(self):
-        return self.session.query(Bug).filter(~Bug.builds.any())
-
-    # returns uninjected (not yet in the build table) possibly fake bugs
-    def uninjected2(self, fake, allowed_bugtypes=None):
-        ret = self.uninjected() \
-            .join(Bug.atp) \
-            .join(Bug.trigger) \
-            .join(DuaBytes.dua) \
-            .filter(Dua.fake_dua == fake)
-        if allowed_bugtypes:
-            ret.filter(Bug.type.in_(allowed_bugtypes))
-        return ret
-
-    def uninjected_random(self, fake, allowed_bugtypes=None):
-        return self.uninjected2(fake, allowed_bugtypes).order_by(func.random())
-
-    def uninjected_random_by_atp_bugtype(self, fake, atp_types=None, allowed_bugtypes=None, atp_lim=10):
-        # For each ATP find X possible bugs,
-        # Returns dict list of lists:
-        #   {bugtype1:[[atp0_bug0, atp0_bug1,..], [atp1_bug0, atp1_bug1,..]],
-        #    bugtype2:[[atp0_bug0, atp0_bug1,..], [atp1_bug0, atp1_bug1,..]]}
-        # Where sublists are randomly sorted
-        if atp_types:
-            _atps = self.session.query(AttackPoint.id).filter(AttackPoint.typ.in_(atp_types)).all()
-        else:
-            _atps = self.session.query(AttackPoint.id).all()
-
-        atps = [r.id for r in _atps]
-        # print(atps)
-        print("Found {} distinct ATPs".format(len(atps)))
-
-        results = {}
-        assert (len(allowed_bugtypes)), "Requires bugtypes"
-
-        for bugtype in allowed_bugtypes:
-            results[bugtype] = []
-            for atp in atps:
-                q = self.session.query(Bug).filter(Bug.atp_id == atp).filter(~Bug.builds.any()) \
-                    .filter(Bug.type == bugtype) \
-                    .join(Bug.atp) \
-                    .join(Bug.trigger) \
-                    .join(DuaBytes.dua) \
-                    .filter(Dua.fake_dua == fake)
-
-                results[bugtype].append(q.order_by(func.random()).limit(atp_lim).all())
-        return results
-
-    def uninjected_random_by_atp(self, fake, atp_types=None,
-                                 allowed_bugtypes=None, atp_lim=10):
-        # For each ATP find X possible bugs,
-        # Returns list of lists: [[atp0_bug0, atp0_bug1,..],
-        # [atp1_bug0, atp1_bug1,..]]
-        # Where sublists are randomly sorted
-        if atp_types:
-            _atps = self.session.query(AttackPoint.id) \
-                .filter(AttackPoint.typ.in_(atp_types)).all()
-        else:
-            _atps = self.session.query(AttackPoint.id).all()
-
-        atps = [r.id for r in _atps]
-        # print(atps)
-        print("Found {} distinct ATPs".format(len(atps)))
-
-        results = []
-        for atp in atps:
-            q = self.session.query(Bug).filter(Bug.atp_id == atp) \
-                .filter(~Bug.builds.any()) \
-                .join(Bug.atp) \
-                .join(Bug.trigger) \
-                .join(DuaBytes.dua) \
-                .filter(Dua.fake_dua == fake)
-            if allowed_bugtypes:
-                q = q.filter(Bug.type.in_(allowed_bugtypes))
-
-            results.append(q.order_by(func.random()).limit(atp_lim).all())
-        return results
-
-    def uninjected_random_limit(self, allowed_bugtypes=None, count=100):
-        # Fast, doesn't support fake bugs, only return IDs of allowed bugtypes
-        ret = self.session.query(Bug) \
-            .filter(~Bug.builds.any()) \
-            .options(load_only(Bug.id))
-        if allowed_bugtypes:
-            ret = ret.filter(Bug.type.in_(allowed_bugtypes))
-        return ret.order_by(func.random()).limit(count).all()
-
-    def uninjected_random_y(self, fake, allowed_bugtypes=None, yield_count=100):
-        # Same as above but yield results
-        ret = self.session.query(Bug) \
-            .filter(~Bug.builds.any()).yield_per(yield_count) \
-            .join(Bug.atp) \
-            .join(Bug.trigger) \
-            .join(DuaBytes.dua) \
-            .filter(Dua.fake_dua == fake)
-        if allowed_bugtypes:
-            ret = ret.filter(Bug.type.in_(allowed_bugtypes))
-        yield ret.all()  # TODO randomize- or is it randomized already?
-
-    def uninjected_random_balance(self, fake, num_required, bug_types):
-        bugs = []
-        types_present = self.session.query(Bug.type) \
-            .filter(~Bug.builds.any()) \
-            .group_by(Bug.type)
-        num_avail = 0
-        for (i,) in types_present:
-            if i in bug_types:
-                num_avail += 1
-        print("%d bugs available of allowed types" % num_avail)
-        assert (num_avail > 0)
-        num_per = num_required / num_avail
-        for (i,) in types_present:
-            if (i in bug_types):
-                bug_query = self.uninjected_random(fake).filter(Bug.type == i)
-                print("found %d bugs of type %d" % (bug_query.count(), i))
-                bugs.extend(bug_query[:num_per])
-        return bugs
-
-    def next_bug_random(self, fake):
-        count = self.uninjected2(fake).count()
-        return self.uninjected2(fake)[random.randrange(0, count)]
 
 
 def run_cmd(cmd, envv=None, timeout=30, cwd=None, rr=False, shell=False):
